monitoring/self-healing/requirements.txt
google-cloud-run==0.10.3
google-cloud-monitoring==2.15.1
google-cloud-logging==3.8.0
requests==2.31.0
functions-framework==3.5.0
